+++ src/Sensorario/Resources/Configurator.php
@@ -8,6 +8,8 @@ final class Configurator
 
     private $resourceName;
 
+    private $ranges;
+
     public function __construct(
         $resourceName,
         Container $container
@@ -64,5 +66,17 @@ final class Configurator
     public function rewrites()
     {
         return $this->container->rewrites();
+    }
+
+    public function ranges()
+    {
+        return $this->container->ranges(
+            $this->resourceName
+        );
+    }
+
+    public function globals()
+    {
+        return $this->container->globals();
     }
 }