+++ test/Sensorario/Resources/ResourceTest.php
@@ -536,7 +536,6 @@ final class ResourceTest extends PHPUnit_Framework_TestCase
                ],
            ])
        );
-
        $properties = [
            'width'  => 3000,
            'height' => 400,
@@ -560,6 +559,80 @@ final class ResourceTest extends PHPUnit_Framework_TestCase
        $this->assertEquals(
            $overWrittenBox,
            $box
+       );
+   }
+
+   /**
+    * @expectedException              RuntimeException
+    * @expectedExceptionMessageRegExp #Value `.*` is out of range: `.*`.#
+    */
+   public function testAcceptRangeOfValues()
+   {
+       $configurator = new Configurator(
+           'foo',
+           new Container([
+               'resources' => [
+                   'foo' => [
+                       'constraints' => [
+                           'allowedRanges' => [
+                               'age' => [
+                                    'more_than' => 3,
+                                    'less_than' => 42,
+                               ],
+                           ],
+                           'allowed' => [
+                               'age'
+                           ],
+                       ],
+                   ],
+               ],
+           ])
+       );
+
+       Resource::box(
+           [ 'age' => 2 ],
+           $configurator
+       );
+   }
+
+   public function testAllResourcesInheritGlobalAllowingConfiguration()
+   {
+       $configurator = new Configurator(
+           'foo',
+           new Container([
+               'globals' => [
+                   'allowed' => [
+                       'width',
+                       'height',
+                   ],
+               ],
+               'resources' => [
+                   'foo' => [
+                       'constraints' => [
+                           'allowed' => [
+                               'foo_size',
+                           ],
+                       ],
+                   ], 
+                   'bar' => [
+                       'constraints' => [
+                           'allowed' => [
+                               'bar_size',
+                           ],
+                       ],
+                   ], 
+               ],
+           ])
+       );
+
+       $resource = Resource::box(
+           [],
+           $configurator
+       );
+
+       $this->assertEquals(
+           ['foo_size', 'width', 'height'],
+           $resource->allowed()
        );
    }
 }